+++ src/meta/assessment/description/analysisAndProcessingDescription.ts
@@ -1,0 +1,4 @@
+export interface AnalysisAndProcessingDescription {
+  estimationAndForecasting: boolean
+  reclassification: boolean
+}
