+++ src/meta/assessment/description/dataSourceColumn.ts
@@ -1,0 +1,8 @@
+export type DataSourceColumn =
+  | 'referenceToTataSource'
+  | 'typeOfDataSource'
+  | 'typeOfDataSourceText'
+  | 'fraVariable'
+  | 'variable'
+  | 'yearForDataSource'
+  | 'comments'
