+++ src/meta/assessment/description/dataSourceVariables.ts
@@ -1,0 +1,7 @@
+import { Label, VariableName } from '@meta/assessment'
+
+export interface DataSourceVariables {
+  // Custom i18n keys to include
+  include: Array<string>
+  prefixes?: Record<VariableName, Label>
+}
