+++ src/meta/assessment/description/nationalDataDescription.ts
@@ -1,0 +1,7 @@
+import { NationalDataDataSourceDescription } from '@meta/assessment/description/nationalDataDataSourceDescription'
+
+export interface NationalDataDescription {
+  dataSources?: NationalDataDataSourceDescription
+  nationalClassification?: boolean
+  originalData?: boolean
+}
