+++ src/meta/assessment/index.ts
@@ -14,7 +14,8 @@ export { Cols } from './cols'
 export type { CommentableDescription, CommentableDescriptionValue, DataSource } from './commentableDescription'
 export { CommentableDescriptionName } from './commentableDescription'
 export type { Cycle, CycledPropsObject, CycleUuid } from './cycle'
-export type { Description, Descriptions } from './description'
+export type { Descriptions } from './description'
+export type { Description } from './description'
 export { dataSourceType } from './descriptions'
 export type { Label } from './label'
 export { Labels } from './labels'
